chore: pr review fixup
