+++ packages/compass-preferences-model/src/index.ts
@@ -25,7 +25,7 @@ export {
 export type { ParsedGlobalPreferencesResult } from './global-config';
 export { getActiveUser, isAIFeatureEnabled } from './utils';
 export { setupPreferencesAndUser } from './compass-utils';
-export type { User, UserStorage } from './storage';
+export type { User, UserStorage } from './user-storage';
 export type { PreferencesAccess };
 export { setupPreferences };
 export const defaultPreferencesInstance: PreferencesAccess =
