+++ packages/compass-preferences-model/src/preferences-storage.ts
@@ -1,0 +1,7 @@
+import type { StoredPreferences } from './preferences-schema';
+
+export interface PreferencesStorage {
+  setup(): Promise<void>;
+  getPreferences(): StoredPreferences;
+  updatePreferences(attributes: Partial<StoredPreferences>): Promise<void>;
+}
